requirements.txt
hydra-core==1.2.0
matplotlib==3.5.2
omegaconf==2.2.2
packaging==21.3
pytest==7.1.2
python-dotenv==0.20.0
pytorch_lightning==1.6.4
rich==12.4.4
scikit_learn==1.1.1
seaborn==0.11.2
setuptools==58.3.0
sh==1.14.2
timm==0.5.4
torch==1.13.1
torchmetrics==0.9.2
torchvision>=0.11.0
hydra-core>=1.1.0
hydra-colorlog>=1.1.0
hydra-optuna-sweeper>=1.1.0
wandb
transformers
nonechucks
git+https://github.com/openai/CLIP.git
